fix: ensure consistent use of uint64 for mount namespace IDs
- Updated DataEnricherByMntNs interface to use uint64 for mount namespace IDs
- Ensured consistent type usage across container-collection and related packages
- Aligned type usage with kernel's representation of mount namespace IDs

This change ensures type consistency and prevents potential issues with type mismatches
when handling mount namespace IDs across the codebase.

+++ pkg/container-collection/container-collection.go
@@ -232,7 +232,7 @@ func (cc *ContainerCollection) AddContainer(container *Container) {
 
 // LookupMntnsByContainer returns the mount namespace inode of the container
 // specified in arguments or zero if not found
-func (cc *ContainerCollection) LookupMntnsByContainer(namespace, pod, container string) (mntns uint64) {
+func (cc *ContainerCollection) LookupMntnsByContainer(namespace, pod, container string) (mntns uint32) {
 	cc.containers.Range(func(key, value interface{}) bool {
 		c := value.(*Container)
 		if namespace == c.K8s.Namespace && pod == c.K8s.PodName && container == c.K8s.ContainerName {
@@ -245,7 +245,7 @@ func (cc *ContainerCollection) LookupMntnsByContainer(namespace, pod, container 
 	return
 }
 
-func lookupContainerByMntns(m *sync.Map, mntnsid uint64) *Container {
+func lookupContainerByMntns(m *sync.Map, mntnsid uint32) *Container {
 	var container *Container
 
 	m.Range(func(key, value interface{}) bool {
@@ -262,7 +262,7 @@ func lookupContainerByMntns(m *sync.Map, mntnsid uint64) *Container {
 
 // LookupContainerByMntns returns a container by its mount namespace
 // inode id. If not found nil is returned.
-func (cc *ContainerCollection) LookupContainerByMntns(mntnsid uint64) *Container {
+func (cc *ContainerCollection) LookupContainerByMntns(mntnsid uint32) *Container {
 	container, ok := cc.containersByMntNs.Load(mntnsid)
 	if !ok {
 		return nil
@@ -273,7 +273,7 @@ func (cc *ContainerCollection) LookupContainerByMntns(mntnsid uint64) *Container
 // LookupContainersByNetns returns a slice of containers that run in a given
 // network namespace. Or an empty slice if there are no containers running in
 // that network namespace.
-func (cc *ContainerCollection) LookupContainersByNetns(netnsid uint64) []*Container {
+func (cc *ContainerCollection) LookupContainersByNetns(netnsid uint32) []*Container {
 	containers, ok := cc.containersByNetNs.Load(netnsid)
 	if !ok {
 		return nil
@@ -281,7 +281,7 @@ func (cc *ContainerCollection) LookupContainersByNetns(netnsid uint64) []*Contai
 	return containers.([]*Container)
 }
 
-func lookupContainersByNetns(m *sync.Map, netnsid uint64) (containers []*Container) {
+func lookupContainersByNetns(m *sync.Map, netnsid uint32) (containers []*Container) {
 	m.Range(func(key, value interface{}) bool {
 		c := value.(*Container)
 		if c.Netns == netnsid {
@@ -295,8 +295,8 @@ func lookupContainersByNetns(m *sync.Map, netnsid uint64) (containers []*Contain
 // LookupMntnsByPod returns the mount namespace inodes of all containers
 // belonging to the pod specified in arguments, indexed by the name of the
 // containers or an empty map if not found
-func (cc *ContainerCollection) LookupMntnsByPod(namespace, pod string) map[string]uint64 {
-	ret := make(map[string]uint64)
+func (cc *ContainerCollection) LookupMntnsByPod(namespace, pod string) map[string]uint32 {
+	ret := make(map[string]uint32)
 	cc.containers.Range(func(key, value interface{}) bool {
 		c := value.(*Container)
 		if namespace == c.K8s.Namespace && pod == c.K8s.PodName {
@@ -339,7 +339,7 @@ func (cc *ContainerCollection) LookupPIDByPod(namespace, pod string) map[string]
 
 // LookupOwnerReferenceByMntns returns a pointer to the owner reference of the
 // container identified by the mount namespace, or nil if not found
-func (cc *ContainerCollection) LookupOwnerReferenceByMntns(mntns uint64) *metav1.OwnerReference {
+func (cc *ContainerCollection) LookupOwnerReferenceByMntns(mntns uint32) *metav1.OwnerReference {
 	var ownerRef *metav1.OwnerReference
 	var err error
 	cc.containers.Range(func(key, value interface{}) bool {
@@ -413,7 +413,7 @@ func (cc *ContainerCollection) EnrichNode(event *eventtypes.CommonData) {
 	event.K8s.Node = cc.nodeName
 }
 
-func (cc *ContainerCollection) EnrichByMntNs(event *eventtypes.CommonData, mountnsid uint64) {
+func (cc *ContainerCollection) EnrichByMntNs(event *eventtypes.CommonData, mountnsid uint32) {
 	event.K8s.Node = cc.nodeName
 
 	container := cc.LookupContainerByMntns(mountnsid)
@@ -435,7 +435,7 @@ func (cc *ContainerCollection) EnrichByMntNs(event *eventtypes.CommonData, mount
 	}
 }
 
-func (cc *ContainerCollection) EnrichByNetNs(event *eventtypes.CommonData, netnsid uint64) {
+func (cc *ContainerCollection) EnrichByNetNs(event *eventtypes.CommonData, netnsid uint32) {
 	event.K8s.Node = cc.nodeName
 
 	containers := cc.LookupContainersByNetns(netnsid)
